ch(db) - Refactor db wrapper.
- Create an abstract Db class.
- Create DbWrapper class.
- Implement all db operations.
- Modify auth and product views to use new methods.
