+++ .env_example
@@ -1,0 +1,5 @@
+DB_NAME=database_name
+DB_USER=database_user
+DB_PASSWORD=database_password
+DB_HOST=db_host
+DB_PORT=db_port
