+++ app/views/auth.py
@@ -3,14 +3,16 @@ from collections import namedtuple
 from datetime import datetime
 
 from flask import Blueprint, jsonify, request
+from flask_jwt_extended import create_access_token
 from werkzeug.security import check_password_hash, generate_password_hash
 
-from db.connection import Connection
-from flask_jwt_extended import create_access_token
-from models.user_model import UserModel
+from connection import Connection
+from db.db_wrapper import DBWrapper
+from app.views.utils import create_user_dict
 
 bp = Blueprint('auth', __name__, url_prefix='/auth')
 db_conn = Connection.instance()
+user_table = DBWrapper(db_conn, 'user_table')
 
 
 @bp.route('/signup', methods=['POST'])
@@ -31,7 +33,6 @@ def signup():
         password=generate_password_hash(password),
         date_created=str(datetime.now())
     )
-    user_table = UserModel(db_conn)
     user_table.create(new_user)
     return jsonify(
         message="Registration Successful"
@@ -42,10 +43,12 @@ def signup():
 def login():
     """logging in an existing user"""
     request_data = request.get_json()
-    email = request_data['email']
+    email_address = request_data['email']
     password = request_data['password']
-    user = UserModel(db_conn)
-    user_dict = user.user_dict(email)
+    Columns = namedtuple('Columns', 'email')
+    user_details = Columns(email=email_address)
+    user_tuple = user_table.select(user_details)
+    user_dict = create_user_dict(user_tuple)
     if (user_dict is not None) and check_password_hash(user_dict['password'], password):
         access_token = create_access_token(identity=user_dict['email'])
         return jsonify(
