+++ app/views/product.py
@@ -4,11 +4,13 @@ from datetime import datetime
 
 from flask import Blueprint, jsonify, request
 
-from db.connection import Connection
-from models.product_model import ProductModel
+from connection import Connection
+from db.db_wrapper import DBWrapper
+from app.views.utils import create_product_dict
 
 bp = Blueprint('product', __name__, url_prefix='/product')
 db_conn = Connection.instance()
+product_table = DBWrapper(db_conn, 'product')
 
 
 @bp.route('/', methods=['POST'])
@@ -24,7 +26,6 @@ def create():
         price=price,
         date_created=str(datetime.now())
     )
-    product_table = ProductModel(db_conn)
     product_table.create(new_product)
     return jsonify(
         message="Product Successful"
@@ -34,32 +35,40 @@ def create():
 @bp.route('/', methods=['GET'])
 def all_products():
     """Get all products"""
-    product_table = ProductModel(db_conn)
-    all_products = product_table.find_all()
+    Columns = namedtuple('Columns', 'id name price date_created')
+    product_list = []
+    for product in product_table.select_all(Columns):
+        product_list.append({
+            'id': product[0],
+            'name': product[1],
+            'price': product[2],
+            'date_created': product[3]
+        })
     return jsonify(
-        products=all_products
+        products=product_list
     ), 200
 
 
 @bp.route('/<int:product_id>', methods=['GET'])
 def single_product(product_id):
     """return a single product given its id"""
-    product = ProductModel(db_conn)
-    selected_product = product.find(product_id)
-    if selected_product is None:
+    Columns = namedtuple('Columns', 'id')
+    single_product = Columns(id=product_id)
+    selected_product = product_table.select(single_product)
+    product_dict = create_product_dict(selected_product)
+    if product_dict is None:
         return jsonify(message="Product not found"), 404
     return jsonify(
-        product=selected_product
+        product=product_dict
     ), 200
 
 
 @bp.route('/<int:product_id>', methods=['DELETE'])
 def delete_product(product_id):
     """delete product from system"""
-    product = ProductModel(db_conn)
     Columns = namedtuple('Columns', 'id')
     del_product = Columns(id=product_id)
-    product.delete(del_product)
+    product_table.delete(del_product)
     return jsonify(
         message=f"Product {product_id} has been deleted"
     ), 200
@@ -71,10 +80,9 @@ def update_product(product_id):
     request_data = request.get_json()
     name = request_data['name']
     price = request_data['price']
-    product = ProductModel(db_conn)
     Row_update = namedtuple('Row_Update', 'id name price')
     product_update = Row_update(id=product_id, name=name, price=price)
-    product.update(product_update)
+    product_table.update(product_update)
     return jsonify(
         message=f"Product {product_id} has been updated"
     ), 200
