--- db/insert.py
@@ -1,24 +1,0 @@
-from db.operation import Operation
-from db.utils import convertTuple, return_values
-
-
-class Insert(Operation):
-    """Handles creating a new record in the db"""
-
-    name = 'Insert'
-
-    def __init__(self, db_conn, table_name, cols_values):
-        super().__init__(db_conn, table_name, cols_values)
-
-    def execute(self):
-        """Insert new record into database"""
-        columns = convertTuple(self.cols_values._fields)
-        values = [
-            getattr(self.cols_values, x) for x in self.cols_values._fields
-            ]
-        str_values = return_values(values)
-        self.db_conn.cur.execute(f"""
-        INSERT INTO {self.table_name} ({columns})
-        VALUES ({str_values})
-        """)
-        self.db_conn.commit_session()
